Refactor: Create BaseCamarasempapelSpider and derive es-linhares and sp-sao-jose-dos-campos
